refactor: consolidate new implementation into State.as method and remove as2
