+++ packages/react/src/jsx-runtime.ts
@@ -1,8 +1,11 @@
-import State from '@expressive/mvc';
+import { State, Context, watch, METHOD } from '@expressive/mvc';
+
 import Runtime from 'react/jsx-runtime';
 import React from 'react';
 
-import { Render, AsComponent, Props } from './state';
+import { AsComponent, Pragma, Props } from './state';
+import { HasProps, RenderProps } from './component';
+import { provide } from './context';
 
 export declare namespace JSX {
   type ElementType =
@@ -49,5 +52,62 @@ export function patch(
 
 export const jsx = patch.bind(Runtime.jsx);
 export const jsxs = patch.bind(Runtime.jsxs);
+
+function Render<T extends AsComponent>(
+  this: State.Type<T>,
+  props: Props<T>,
+  props2?: Props<T>
+) {
+  const { is, ...rest } = { ...props, ...props2 };
+
+  const ambient = Context.use();
+  const state = Pragma.useState<(props: any) => any>(() => {
+    const instance = this.new(rest as {}, is && ((x) => void is(x)));
+    const context = ambient.push(instance);
+
+    let ready: boolean | undefined;
+    let active: T;
+
+    watch(instance, (current) => {
+      active = current;
+
+      if (ready) state[1]((x) => x.bind(null));
+    });
+
+    function didMount() {
+      ready = true;
+      return () => {
+        context.pop();
+        instance.set(null);
+      };
+    }
+
+    function Render(props: Props<T>) {
+      const render = METHOD.get(active.render) || props.render || active.render;
+
+      return render
+        ? render.call(active, props as HasProps<T>, active)
+        : props.children;
+    }
+
+    return (props: RenderProps<T>) => {
+      ready = false;
+
+      Pragma.useEffect(didMount, []);
+      Promise.resolve(instance.set(props as {})).finally(() => {
+        ready = true;
+      });
+
+      return provide(
+        context,
+        Pragma.createElement(Render, props as any),
+        props.fallback || active.fallback,
+        String(instance)
+      );
+    };
+  });
+
+  return state[0](rest);
+}
 
 export { Fragment } from 'react';
